Split utils to common and render parts
